refactor: move format enum to parameters package
