+++ eno-core/src/main/java/fr/insee/eno/core/model/calculated/CalculatedExpression.java
@@ -3,7 +3,7 @@ package fr.insee.eno.core.model.calculated;
 import datacollection33.IfThenElseType;
 import fr.insee.eno.core.Constant;
 import fr.insee.eno.core.annotations.DDI;
-import fr.insee.eno.core.annotations.Format;
+import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.annotations.Lunatic;
 import fr.insee.eno.core.model.EnoObject;
 import fr.insee.lunatic.model.flat.LabelType;
