+++ eno-core/src/main/java/fr/insee/eno/core/model/label/Label.java
@@ -2,7 +2,7 @@ package fr.insee.eno.core.model.label;
 
 import fr.insee.eno.core.Constant;
 import fr.insee.eno.core.annotations.DDI;
-import fr.insee.eno.core.annotations.Format;
+import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.annotations.Lunatic;
 import fr.insee.eno.core.model.EnoObject;
 import lombok.Getter;
