+++ eno-core/src/main/java/fr/insee/eno/core/model/question/TableCell.java
@@ -3,7 +3,7 @@ package fr.insee.eno.core.model.question;
 import datacollection33.GridResponseDomainInMixedType;
 import datacollection33.QuestionItemType;
 import fr.insee.eno.core.annotations.DDI;
-import fr.insee.eno.core.annotations.Format;
+import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.annotations.Lunatic;
 import fr.insee.eno.core.model.EnoObject;
 import fr.insee.eno.core.model.code.CodeItem;
