+++ eno-core/src/main/java/fr/insee/eno/core/processing/EnoProcessing.java
@@ -1,6 +1,6 @@
 package fr.insee.eno.core.processing;
 
-import fr.insee.eno.core.annotations.Format;
+import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.model.EnoQuestionnaire;
 import fr.insee.eno.core.parameter.EnoParameters;
 import fr.insee.eno.core.processing.impl.*;
