+++ eno-core/src/test/java/fr/insee/eno/core/DDIToEnoTest.java
@@ -1,6 +1,6 @@
 package fr.insee.eno.core;
 
-import fr.insee.eno.core.annotations.Format;
+import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.mappers.DDIMapper;
 import fr.insee.eno.core.model.EnoQuestionnaire;
 import fr.insee.eno.core.parsers.DDIParser;
