+++ eno-ws/src/main/java/fr/insee/eno/ws/controller/V3ParameterController.java
@@ -1,6 +1,6 @@
 package fr.insee.eno.ws.controller;
 
-import fr.insee.eno.core.annotations.Format;
+import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.parameter.EnoParameters;
 import fr.insee.eno.ws.controller.utils.HeadersUtils;
 import fr.insee.eno.ws.service.ParameterService;
